requirements.txt
autopep8==2.0.4
blinker==1.7.0
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycodestyle==2.11.1
python-dotenv==1.0.0
Werkzeug==3.0.1
